# RecruitmentAgency/RecruitmentAgency.WASM/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RecruitmentAgency.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Position": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Возвращает список всех позиций.",
        "responses": {
          "200": {
            "description": "Список успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Добавляет новую позицию.",
        "requestBody": {
          "description": "Информация о новой позиции.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Позиция успешно добавлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              }
            }
          },
          "400": {
            "description": "Позиция не была добавлена."
          }
        }
      },
      "put": {
        "tags": [
          "Position"
        ],
        "summary": "Обновляет информацию о существующей позиции.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор должности",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновлённая информация о позиции.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Позиция успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              }
            }
          },
          "404": {
            "description": "Позиция с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/Position/{id}": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Получает информацию о позиции по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор позиции.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Позиция найдена и возвращена успешно.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDTO"
                }
              }
            }
          },
          "404": {
            "description": "Позиция с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "Position"
        ],
        "summary": "Удаляет позицию по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор позиции.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Позиция успешно удалена."
          },
          "404": {
            "description": "Позиция с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/Query/applicants/position/{positionName}": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает всех соискателей для указанной позиции, отсортированных по полному имени.",
        "parameters": [
          {
            "name": "positionName",
            "in": "path",
            "description": "Название позиции.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список соискателей успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Query/applicants/date-range": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает всех соискателей в диапазоне дат подачи заявок.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Дата начала диапазона.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Дата конца диапазона.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список соискателей успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Query/applicants/employer-application/{employerApplicationId}": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает соискателей для заявки работодателя по идентификатору заявки.",
        "parameters": [
          {
            "name": "employerApplicationId",
            "in": "path",
            "description": "Идентификатор заявки работодателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список соискателей успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Query/employers/max-salary": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает работодателей с максимальными зарплатами по заявкам.",
        "responses": {
          "200": {
            "description": "Список работодателей успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Список работодателей отсутствует."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApplicantDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор соискателя",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "description": "ФИО соискателя",
            "nullable": true
          },
          "contactInformation": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "experience": {
            "type": "number",
            "description": "Опыт работы",
            "format": "double"
          },
          "education": {
            "type": "string",
            "description": "Образование",
            "nullable": true
          },
          "salaries": {
            "type": "integer",
            "description": "Ожидаемый уровень зарплаты",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO Соискателя"
      },
      "ApplicantsForEmployerApplicationDTO": {
        "type": "object",
        "properties": {
          "applicantId": {
            "type": "integer",
            "description": "Соискатель",
            "format": "int32"
          },
          "employerAppId": {
            "type": "integer",
            "description": "Заявка работодателя",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO описывающее соискателя и заявку работодателя"
      },
...
      "EmployerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор работадателя",
            "format": "int32"
          },
          "companyName": {
            "type": "string",
            "description": "Название компании",
            "nullable": true
          },
          "contactPersonName": {
            "type": "string",
            "description": "ФИО контактного лица",
            "nullable": true
          },
          "companyNumber": {
            "type": "string",
            "description": "Телефон работодателя",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO работодателя"
      },
      "PositionCreateDTO": {
        "type": "object",
        "properties": {
          "section": {
            "type": "string",
            "description": "Раздел",
            "nullable": true
          },
          "positionName": {
            "type": "string",
            "description": "Должность",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания должности"
      },
      "PositionDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор должности",
            "format": "int32"
          },
          "section": {
            "type": "string",
            "description": "Раздел",
            "nullable": true
          },
          "positionName": {
            "type": "string",
            "description": "Должность",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO должности"
      },
...
    }
  }
}